Hma 10250 (#176)
* HMA-10250 unit test cases fixed

* HMA-10250 fix it test cases

* HMA-10250 fix it test cases and coverage

* HMA-10250 Prr bot coments addressed

* HMA-10250 PR comments addressed

* HMA-10250 remove print stmt

+++ app/uk/gov/hmrc/mobilepaye/config/MobilePayeConfig.scala
@@ -22,5 +22,5 @@ case class MobilePayeConfig @Inject() (configuration: Configuration) {
 
   val internalAuthResourceType: String = configuration.get[String]("microservice.services.internal-auth.resource-type")
 
-  val mongoTtl: Int = configuration.get[Int]("mongodb.ttlSecond")
+  val mongoTtl: Long = configuration.get[Long]("mongodb.ttlSecond")
 }
